# backend/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: MockVision API
  version: 0.1.0
  description: |
    REST API of a MockVision node, shared by the panel (session cookie) and
    automation (API tokens sent as `Authorization: Bearer mvt_…`). Errors use
    problem+json (RFC 9457). State-changing requests with the session cookie
    must send the `X-MockVision-Request` header; requests with a token need
    not. A token with the read scope may only make GET requests and open the
    WebSocket; tokens are created and revoked from the panel only. Live
    updates come through a single WebSocket at `/api/v1/ws` (see the
    `x-websocket` section).
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
security:
  - session: []
  - token: []
...
paths:
  /auth/setup:
    post:
      tags: [auth]
      operationId: setup
      summary: Create the first administrator (first run only)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "201": { description: Created and logged in, content: { application/json: { schema: { $ref: "#/components/schemas/Me" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200": { description: Logged in, content: { application/json: { schema: { $ref: "#/components/schemas/Me" } } } }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
...
  /auth/me:
    get:
      tags: [auth]
      operationId: me
      summary: Current user, and the token of a request made with one; 401 carries setup_required on a fresh node
      responses:
        "200": { description: Session, content: { application/json: { schema: { $ref: "#/components/schemas/Me" } } } }
        "401": { $ref: "#/components/responses/Problem" }
...
  /node/history:
    get:
      tags: [node]
      operationId: getNodeHistory
      summary: Node samples of the last ten minutes, every 2 s
      parameters:
        - { name: since, in: query, description: Unix milliseconds or RFC 3339, schema: { type: string } }
      responses:
        "200":
          description: Samples, oldest first
          content:
            application/json:
              schema:
                type: object
                required: [samples]
                properties:
                  samples: { type: array, items: { $ref: "#/components/schemas/NodeSample" } }
  /settings:
    get:
      tags: [node]
      operationId: getSettings
      responses:
        "200": { description: Settings, content: { application/json: { schema: { $ref: "#/components/schemas/Settings" } } } }
    patch:
      tags: [node]
      operationId: patchSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsPatch" }
      responses:
        "200": { description: Settings, content: { application/json: { schema: { $ref: "#/components/schemas/Settings" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /packages:
    post:
      tags: [profiles]
      operationId: importPackage
      summary: Import a .mvpkg package or a loose profile.yaml (draft)
      description: |
        The import runs as a job. The call waits up to a minute for it and
        answers with the result; when the job is still queued or running it
        answers 202 with the job, to follow at /jobs/{id}.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200": { description: Already installed, content: { application/json: { schema: { $ref: "#/components/schemas/ImportResult" } } } }
        "201": { description: Installed, content: { application/json: { schema: { $ref: "#/components/schemas/ImportResult" } } } }
        "202":
          description: The import goes on as a job
          content:
            application/json:
              schema:
                type: object
                required: [job]
                properties:
                  job: { $ref: "#/components/schemas/Job" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /profiles/{vendor}/{model}/versions/{version}:
    parameters:
      - { name: vendor, in: path, required: true, schema: { type: string } }
      - { name: model, in: path, required: true, schema: { type: string } }
      - { name: version, in: path, required: true, schema: { type: string } }
    get:
      tags: [profiles]
      operationId: getProfile
      responses:
        "200": { description: Profile, content: { application/json: { schema: { $ref: "#/components/schemas/ProfileDetail" } } } }
        "404": { $ref: "#/components/responses/Problem" }
  /profiles/{vendor}/{model}/versions/{version}/actions/{action}:
    parameters:
      - { name: vendor, in: path, required: true, schema: { type: string } }
      - { name: model, in: path, required: true, schema: { type: string } }
      - { name: version, in: path, required: true, schema: { type: string } }
      - { name: action, in: path, required: true, schema: { type: string, enum: [archive, unarchive] } }
    post:
      tags: [profiles]
      operationId: profileAction
      responses:
        "200": { description: Profile, content: { application/json: { schema: { $ref: "#/components/schemas/ProfileDetail" } } } }
  /tokens:
    get:
      tags: [tokens]
      operationId: listTokens
      summary: API tokens of the current user (panel session only)
      responses:
        "200":
          description: Tokens, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Token" } }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [tokens]
      operationId: createToken
      summary: Create an API token (panel session only); the secret is returned once
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenInput" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/CreatedToken" } } } }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /tokens/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    delete:
      tags: [tokens]
      operationId: revokeToken
      summary: Revoke an API token (panel session only)
      responses:
        "204": { description: Revoked }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /jobs:
    get:
      tags: [jobs]
      operationId: listJobs
      summary: Background jobs, newest first (D72, D74)
      parameters:
        - { name: status, in: query, description: A status, active (queued, running, waiting or interrupted) or finished, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [rendition, import, renditions.prepare] } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200": { description: Jobs, content: { application/json: { schema: { $ref: "#/components/schemas/JobPage" } } } }
        "422": { $ref: "#/components/responses/Problem" }
    post:
      tags: [jobs]
      operationId: createJob
      summary: Start a job; for now renditions.prepare, which encodes what the cameras need
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobInput" }
      responses:
        "202": { description: Queued, or the open job doing the same work, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [jobs]
      operationId: getJob
      summary: A job with its history
      responses:
        "200": { description: Job, content: { application/json: { schema: { $ref: "#/components/schemas/JobDetail" } } } }
        "404": { $ref: "#/components/responses/Problem" }
  /jobs/{id}/actions/{action}:
    parameters:
      - $ref: "#/components/parameters/ID"
      - { name: action, in: path, required: true, schema: { type: string, enum: [cancel, resume, answer] } }
    post:
      tags: [jobs]
      operationId: jobAction
      summary: Cancel a job, resume an interrupted one, or answer its question ({"answer":"<option id>"})
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                answer: { type: string }
      responses:
        "200": { description: Job, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /audit:
    get:
      tags: [audit]
      operationId: listAudit
      summary: Who changed what, when and from where (kept 90 days)
      parameters:
        - { name: origin, in: query, schema: { type: string, enum: [panel, api, camera, system] } }
        - { name: entity_type, in: query, schema: { type: string } }
        - { name: entity_id, in: query, schema: { type: string } }
        - { name: token_id, in: query, schema: { type: string } }
        - { name: action, in: query, description: Exact action or prefix, such as camera, schema: { type: string } }
        - { name: since, in: query, description: Unix milliseconds or RFC 3339, schema: { type: string } }
        - { name: until, in: query, description: Unix milliseconds or RFC 3339, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200": { description: Entries, newest first, content: { application/json: { schema: { $ref: "#/components/schemas/AuditPage" } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras:
    get:
      tags: [cameras]
      operationId: listCameras
      parameters:
        - { name: q, in: query, description: Text in the name, address, MAC, serial or tags, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: profile, in: query, description: Profile ID or ID@version, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
      responses:
        "200":
          description: Cameras
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Camera" } }
    post:
      tags: [cameras]
      operationId: createCamera
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCamera" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/actions/bulk:
    post:
      tags: [cameras]
      operationId: bulkCameras
      summary: Start, stop, restart, clone or delete several cameras; each gets its own result
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkAction" }
      responses:
        "200": { description: Results, content: { application/json: { schema: { $ref: "#/components/schemas/BulkResult" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [cameras]
      operationId: getCamera
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [cameras]
      operationId: updateCamera
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCamera" }
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
    delete:
      tags: [cameras]
      operationId: deleteCamera
      responses:
        "204": { description: Deleted }
  /cameras/{id}/actions/{action}:
    parameters:
      - $ref: "#/components/parameters/ID"
      - { name: action, in: path, required: true, schema: { type: string, enum: [start, stop, restart] } }
    post:
      tags: [cameras]
      operationId: cameraAction
      responses:
        "202": { description: Accepted, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "409": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/actions/factory-reset:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [cameras]
      operationId: resetCamera
      description: >
        Restores the camera to its profile (RN-10): parameters, accounts and
        protocols; with scope full, also the profile's factory address. The
        picture is kept. A running camera reboots.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scope]
              properties:
                scope: { type: string, enum: [settings, full] }
      responses:
        "202": { description: Accepted, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/actions/clone:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [cameras]
      operationId: cloneCamera
      description: >
        Creates a camera with the same profile, parameters, accounts,
        protocols, picture and targets, with its own ID, serial, MAC and
        address.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CloneCamera" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/users:
    parameters:
      - $ref: "#/components/parameters/ID"
    put:
      tags: [cameras]
      operationId: setCameraUsers
      description: >
        Replaces the camera's accounts (RN-11: at least one admin). An empty
        password keeps the account's current one. Applied at once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [users]
              properties:
                users: { type: array, items: { $ref: "#/components/schemas/CameraUserInput" } }
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/protocols:
    parameters:
      - $ref: "#/components/parameters/ID"
    put:
      tags: [cameras]
      operationId: setCameraProtocols
      description: >
        Enables, disables or moves protocols of the camera's profile (RN-04).
        A running camera applies them when it restarts (RN-09).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [protocols]
              properties:
                protocols:
                  type: array
                  items:
                    type: object
                    required: [instance]
                    properties:
                      instance: { type: string }
                      enabled: { type: boolean }
                      port: { type: integer, minimum: 0, maximum: 65535 }
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/streams/{stream}:
    parameters:
      - $ref: "#/components/parameters/ID"
      - { name: stream, in: path, required: true, schema: { type: string } }
    patch:
      tags: [cameras]
      operationId: updateCameraStream
      description: >
        Changes the picture, resolution or frame rate of a stream. The stream
        is encoded again and a running camera switches to it without
        restarting.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                asset_id: { type: string }
                resolution: { type: string }
                fps: { type: integer }
      responses:
        "200": { description: Camera, content: { application/json: { schema: { $ref: "#/components/schemas/Camera" } } } }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /cameras/{id}/status:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [cameras]
      operationId: getCameraStatus
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                type: object
                required: [status, endpoints, metrics]
                properties:
                  status: { $ref: "#/components/schemas/CameraStatus" }
                  endpoints: { type: array, items: { $ref: "#/components/schemas/Endpoint" } }
                  metrics: { oneOf: [{ $ref: "#/components/schemas/CameraMetrics" }, { type: "null" }] }
  /cameras/{id}/config:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [cameras]
      operationId: getCameraConfig
      responses:
        "200": { description: Parameters, content: { application/json: { schema: { $ref: "#/components/schemas/ParamList" } } } }
    patch:
      tags: [cameras]
      operationId: patchCameraConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [values]
              properties:
                values: { type: object, additionalProperties: true }
      responses:
        "200": { description: Parameters, content: { application/json: { schema: { $ref: "#/components/schemas/ParamList" } } } }
        "422": { $ref: "#/components/responses/Problem" }
...
  /cameras/{id}/snapshot:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [cameras]
      operationId: getSnapshot
      responses:
        "200": { description: JPEG, content: { image/jpeg: { schema: { type: string, format: binary } } } }
  /cameras/{id}/events:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [events]
      operationId: triggerEvent
      summary: Manual trigger; the camera serializes and delivers the event
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Trigger" }
      responses:
        "201": { description: Emitted, content: { application/json: { schema: { $ref: "#/components/schemas/Event" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /assets:
    get:
      tags: [assets]
      operationId: listAssets
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Asset" } }
    post:
      tags: [assets]
      operationId: uploadAsset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201": { description: Stored, content: { application/json: { schema: { $ref: "#/components/schemas/Asset" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /assets/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    delete:
      tags: [assets]
      operationId: deleteAsset
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Problem" }
  /assets/{id}/content:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [assets]
      operationId: getAssetContent
      responses:
        "200": { description: Image, content: { image/*: { schema: { type: string, format: binary } } } }
...
  /targets/{id}/actions/test:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [targets]
      operationId: testTarget
      responses:
        "200": { description: Result, content: { application/json: { schema: { $ref: "#/components/schemas/TargetTest" } } } }
...
  /events/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [events]
      operationId: getEvent
      responses:
        "200": { description: Event, content: { application/json: { schema: { $ref: "#/components/schemas/Event" } } } }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: mv_session
    token:
      type: http
      scheme: bearer
      description: API token created in the panel (Settings, API tokens)
...
  responses:
    Problem:
      description: Problem
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Credentials:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    Me:
      type: object
      required: [user, expires_at]
      properties:
        user:
          type: object
          required: [id, username, role]
          properties:
            id: { type: string }
            username: { type: string }
            role: { type: string }
        expires_at: { type: [string, "null"], format: date-time, description: Null for a token that does not expire }
        token:
          type: object
          required: [id, name, scopes]
          properties:
            id: { type: string }
            name: { type: string }
            scopes: { type: array, items: { type: string, enum: [read, write] } }
...